+++ R/mm_check_mcmc_file.R
@@ -59,6 +59,7 @@ mm_check_mcmc_file <- function(model_file) {
 #' \dontrun{
 #' # takes a long time, so run only when needed
 #' checks <- streamMetabolizer:::mm_check_mcmc_files()
+#' saveRDS(checks, file='temp/bayes_model_checks.Rds')
 #' checks <- streamMetabolizer:::mm_check_mcmc_files("\\.jags")
 #' checks <- streamMetabolizer:::mm_check_mcmc_files("*ko\\.stan")
 #' checks <- streamMetabolizer:::mm_check_mcmc_files("b_np_.*_ko\\.stan")